refactor: separa ConfigProvider
Componentes
