+++ src/contexts/ConfigProvider.tsx
@@ -1,0 +1,15 @@
+import { useReducer } from "react";
+import { configReducer, initialConfig } from "./configReducer";
+import { ConfigContext, ConfigDispatchContext } from "./ConfigContext";
+
+export function ConfigProvider({ children }: { children: React.ReactNode }) {
+  const [config, dispatch] = useReducer(configReducer, initialConfig);
+
+  return (
+    <ConfigContext.Provider value={config}>
+      <ConfigDispatchContext.Provider value={dispatch}>
+        {children}
+      </ConfigDispatchContext.Provider>
+    </ConfigContext.Provider>
+  );
+}
